Fix split package in search business rules module
The CappedScore lucene query exists in search-business-rules, but split
the lucene search package. This commit moves the implementation class
into the existing module package name.
